Stukje code birgit beveiligd
git-svn-id: svn://localhost/repo@326 10651035-e252-4840-a10e-5981b4cf34d1

+++ trunk/Hoorcollege/lector/verwerkCommentaarVerwijderen.php
@@ -1,11 +1,12 @@
 <?php
 include_once('./../includes/kern.php');
 session_start();
+if(isset($_SESSION['gebruiker']) && $_SESSION['gebruiker']->getNiveau() >= 40) { //lector is ingelogged
+    $commentaarId = $_GET['id'];
+    echo $commentaarId;
 
-$commentaarId = $_GET['id'];
-echo $commentaarId;
-
-if(validateNumber($commentaarId)){
-    $db->Execute('DELETE FROM hoorcollege_reactie WHERE idReactie='.$commentaarId);
+    if(validateNumber($commentaarId)){
+        $db->Execute('DELETE FROM hoorcollege_reactie WHERE idReactie='.$commentaarId);
+    }
 }
 ?>
